Fix building of kernel module for older kernels.
This driver is intended for use on older kernels, but commit 8471176
breaks this by updating the class_create API to be compatible with newer
kernels.
Add in kernel version check - but it comes with the caveat that this
driver should not be used on kernels newer than 5.17. I've left it
able to build in case there is some strange cornercase I don't know of
where the firmware-attributes class can't be used.

Highlighted the requirements more explicitely in the README's to try and
avoid furthre confusion.

Fixes: 8471176 - Advantest issue fix

Signed-off-by: Mark Pearson <mpearson-lenovo@squebb.ca>